chore: Add .prettierignore file and update README.md
This commit adds a .prettierignore file to the repository, which includes the pattern `**/*.html` to exclude HTML files from Prettier formatting. This was necessary as it was messing up with the Django templates.

Additionally, it updates the README.md file to fix a typo and improve the readability of the description of the EduVerse project.