refactor: Add created_at and updated_at fields to User model
This commit adds the "created_at" and "updated_at" fields to the User model in the core/models.py file. The "created_at" field is a DateTimeField that stores the timestamp of when the user was created, while the "updated_at" field is a DateTimeField that stores the timestamp of the user's last update. This change allows for tracking the creation and update timestamps of a user.