feat: Update dashboard base.html to include institution link for owner
This commit modifies the dashboard_base.html template to include a link to the institution page for the owner of the profile. This link is only displayed if the logged-in user is the owner of the institution. This change improves the user experience by providing easy access to the institution settings for the owner.

Co-authored-by: Agwu Emmanuel <emmaagwu@users.noreply.github.com>