feat: Update user profile and dashboard templates
This commit updates the user_profile_detail.html and dashboard_base.html templates.

In user_profile_detail.html, a user profile update form is added, allowing users to conveniently edit their profile details. This improves the user experience by providing a seamless way for users to update their information.

In dashboard_base.html, a link to the institution page is included for the owner of the profile. This link is only displayed if the logged-in user is the owner of the institution. This change enhances the user experience by providing easy access to the institution settings for the owner.

Co-authored-by: Agwu Emmanuel <emmaagwu@users.noreply.github.com>